pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photorec-sorter"
version = "0.2.1"
authors = [{ name = "RecRanger", email = "RecRanger+package@proton.me" }]
description = "A tool to sort/organize files recovered by the PhotoRec tool"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Operating System :: OS Independent",
]

dependencies = ["loguru", "exifread"]

[dependency-groups]
dev = [
    "pillow>=12.1.0",
    "pyright>=1.1.407",
    "pytest>=8.3.5",
    "ruff>=0.14.10",
]

[project.urls]
Homepage = "https://github.com/RecRanger/photorec-sorter"
Issues = "https://github.com/RecRanger/photorec-sorter/issues"

[project.scripts]
photorec_sorter = "photorec_sorter.cli:main_cli"

[tool.pyright]
typeCheckingMode = "basic"
